requirements.txt
python-telegram-bot==22.0
python-dotenv==1.0.0
yt-dlp==2025.03.31
sqlalchemy==1.4.51  # Versión más estable para Python 3.13
aiosqlite==0.19.0
pydantic==1.10.13
bleach==6.1.0
python-slugify==8.0.1